同步1.5.0 09cc7f597fc187168c6c77ca235d56658c97d8d9     !3453 arange算子默认输出类型和cpu修改一致     Merge pull request !3453 from aikede/b_arange

+++ pytorch1.5.0/src/aten/src/ATen/native/npu/ArangeKernelNpu.cpp
@@ -76,7 +76,7 @@ Tensor arange_npu(
 
   // check start == end
   Tensor result_check = set_to_integral_dtype
-      ? at::empty_with_format({0}, options.dtype(at::ScalarType::Int), ACL_FORMAT_ND)
+      ? at::empty_with_format({0}, options.dtype(at::ScalarType::Long), ACL_FORMAT_ND)
       : at::empty_with_format({0}, options, ACL_FORMAT_ND);
   if (start_value == end_value) {
     return result_check;
@@ -89,7 +89,7 @@ Tensor arange_npu(
   SmallVector<int64_t, SIZE> outputSize = {size_value};
 
   Tensor result = set_to_integral_dtype
-      ? at::empty_with_format(outputSize, options.dtype(at::ScalarType::Int), ACL_FORMAT_ND)
+      ? at::empty_with_format(outputSize, options.dtype(at::ScalarType::Long), ACL_FORMAT_ND)
       : at::empty_with_format(outputSize, options, ACL_FORMAT_ND);
 
   if(options.dtype() == at::kHalf) {
